refactor: 투구가 끝난상태와 진행중인 상태를 분리하여 중복제거
