+++ src/main/java/bowling/domain/score/Pin.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 import bowling.exception.BowlingPinException;
