+++ src/main/java/bowling/domain/state/FirstBowl.java
@@ -1,10 +1,8 @@
 package bowling.domain.state;
 
-import bowling.domain.Pin;
-import bowling.domain.Score;
-import bowling.exception.BowlingStateException;
+import bowling.domain.score.Pin;
 
-public class FirstBowl extends State {
+public class FirstBowl extends Running {
 
     public FirstBowl(int pin) {
         firstPin = new Pin(pin);
@@ -16,15 +14,5 @@ public class FirstBowl extends State {
             return new Spare(firstPin.count(), secondCount);
         }
         return new Miss(firstPin.count(), secondCount);
-    }
-
-    @Override
-    public Score getScore() {
-        throw new BowlingStateException("투구가 완료되지 않아 점수를 확인할 수 없습니다.");
-    }
-
-    @Override
-    public boolean stateFinish() {
-        return false;
     }
 }
