+++ src/main/java/bowling/domain/state/Miss.java
@@ -1,10 +1,9 @@
 package bowling.domain.state;
 
-import bowling.domain.Pin;
-import bowling.domain.Score;
-import bowling.exception.BowlingStateException;
+import bowling.domain.score.Pin;
+import bowling.domain.score.Score;
 
-public class Miss extends State {
+public class Miss extends Finished {
     public Miss(int firstCount, int secondCount) {
         pinValidate(firstCount, secondCount);
         this.firstPin = new Pin(firstCount);
@@ -19,15 +18,5 @@ public class Miss extends State {
     @Override
     public Score getScore() {
         return Score.ofMiss(firstPin.count(), secondPin.count());
-    }
-
-    @Override
-    public State bowl(int pin) {
-        throw new BowlingStateException("다음 프레임에서 투구해주세요.");
-    }
-
-    @Override
-    public boolean stateFinish() {
-        return true;
     }
 }
